allows for selection of single MIDI port
This adds a "(select a MIDI Port)" option to the drop down list of MIDI ports in the event that only one port is available. A bug in the general Companion core does not update a dropdown until the value is changed from the default, so this forces the user to make a change since the default is automatically an invalid selection.